mod server
get custom 404 or default if none in content map
