+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod commands;
 pub mod generators;
 pub mod parsers;
+pub mod server;
