+++ src/server/mod.rs
@@ -1,2 +1,3 @@
 mod server;
-pub use server::*;
+pub use server::start;
+
